requirements.txt
Flask==3.1.0
flask_cors==5.0.1
g4f==0.4.7.7
python-dotenv==1.1.0